Moved a bit of logic to common file
